Impl. `crate::point_src::IdentifiablePartRandPoints::copying_iter`

+++ src/state_serv/simul_params.rs
@@ -3,6 +3,6 @@ use derive_more::Constructor;
 /// Simulation parameters
 #[derive(Constructor, Clone, Copy, Debug)]
 pub struct SimulParams {
-    pub group_count: u32,
-    pub point_count: u32,
+    pub group_count: u8,
+    pub point_count: usize,
 }
